requirements.txt
Flask>=3.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
python-dotenv
PyMuPDF